refactor: cleaned the code based on PR comments
